+++ deepbots/robots/controllers/robot_emitter_receiver.py
@@ -1,79 +1,92 @@
-from abc import abstractmethod
+from abc import ABC, abstractmethod
 
-from .robot_controller import RobotEmitterReceiver
+from controller import Robot
 
 
-class RobotEmitterReceiverCSV(RobotEmitterReceiver):
+class RobotEmitterReceiver(ABC):
     """
-    Basic implementation of a robot that can emit and receive messages to/from the supervisor in string utf-8 form
-    that are comma separated, i.e. a list.
+    This RobotEmitterReceiver implements only the most basic run method, that steps the robot and calls the handleEmitter,
+    handleReceiver methods that are needed for communication with the supervisor.
+
+    This class must be inherited by all robot controllers created by the user and the handleEmitter, handleReceiver,
+    initialize_comms methods are all abstract and need to be implemented, according to their docstrings.
+
+    For a simpler RobotController that implements the methods in a basic form inherit the RobotEmitterReceiver class.
     """
 
     def __init__(
         self, emitter_name="emitter", receiver_name="receiver", timestep=None
     ):
-        super().__init__(timestep=timestep)
+        """
+        The basic robot constructor.
 
+        Initializes the Webots Robot and sets up a basic timestep if None is supplied.
+
+        Also initializes the emitter and the receiver used to communicate with the supervisor, using the
+        initialize_comms() method which must be implemented by the user.
+
+        The two methods handle_emitter() and handle_receiver() must also be implemented by the user.
+
+        For the step argument see relevant Webots documentation:
+        https://cyberbotics.com/doc/guide/controller-programming#the-step-and-wb_robot_step-functions
+
+        :param timestep: float, positive or None
+        """
+        self.robot = Robot()
+
+        if timestep is None:
+            self.timestep = int(self.robot.getBasicTimeStep())
+        else:
+            self.timestep = timestep
+
+        self.emitter, self.receiver = self.initialize_comms(
+            emitter_name, receiver_name
+        )
+
+    def get_timestep(self):
+        return self.timestep
+
+    @abstractmethod
     def initialize_comms(self, emitter_name, receiver_name):
         """
-        This method implements the basic emitter/receiver initialization that assumes that an emitter and a receiver
-        components are present on the Webots robot with appropriate DEFs ("emitter"/"receiver").
-        :return: emitter and receiver references
-        """
+        This method should initialize and the return emitter and receiver in a tuple as expected by the constructor.
+
+        A basic example implementation can be:
+
         emitter = self.robot.getEmitter("emitter")
         receiver = self.robot.getReceiver("receiver")
         receiver.enable(self.timestep)
         return emitter, receiver
 
+        :return: (emitter, receiver) tuple, as initialized
+        """
+        pass
+
+    @abstractmethod
     def handle_emitter(self):
         """
-        This emitter uses the user-implemented create_message() method to get whatever data the robot gathered,
-        convert it to a string if needed and then use the emitter to send the data in a string utf-8 encoding to the
-        supervisor.
+        This method should take data from the robot, eg. sensor data, parse it into a message and use the robot's
+        emitter to send it to the supervisor. This message will be used as the observation of the robot.
         """
-        data = self.create_message()
-        string_message = ""
-        # message can either be a list that needs to be converted in a string or a straight-up string
-        if type(data) is list:
-            string_message = ",".join(map(str, data))
-        elif type(data) is str:
-            string_message = data
-        else:
-            raise TypeError(
-                "message must be either a comma-separated string or a 1D list"
-            )
+        pass
 
-        string_message = string_message.encode("utf-8")
-        self.emitter.send(string_message)
-
+    @abstractmethod
     def handle_receiver(self):
         """
-        This receiver uses the basic Webots receiver-handling code. The use_message_data() method should be implemented
-        to actually use the data received from the supervisor.
-        """
-        if self.receiver.getQueueLength() > 0:
-            # Receive and decode message from supervisor
-            message = self.receiver.getData().decode("utf-8")
-            # Convert string message into a list
-            message = message.split(",")
+        This method should take data through the receiver in the form of a message and parse into data usable by the
+        robot.
 
-            self.use_message_data(message)
-
-            self.receiver.nextPacket()
-
-    @abstractmethod
-    def create_message(self):
-        """
-        This method should be implemented to convert whatever data the robot has, eg. sensor data, into a message to be
-        sent to the supervisor via the emitter.
-        :return: a list or a comma-separated string containing all data
+        For example the message might include a motor speed, which should be parsed and applied to the robot's motor.
         """
         pass
 
-    @abstractmethod
-    def use_message_data(self, message):
+    def run(self):
         """
-        This method should be implemented to apply whatever actions the message (received from the supervisor) contains.
-        :param message: list containing data received from the supervisor
+        This method is required by Webots to update the robot in the simulation. It steps the robot and in each
+        step it runs the two handler methods to use the emitter and receiver components.
+
+        This method should be called by a robot manager to run the robot.
         """
-        pass
+        while self.robot.step(self.timestep) != -1:
+            self.handle_receiver()
+            self.handle_emitter()
